configure: Make arch_libdir a first-class entity
arch_libdir was brought in to improve the heuristic determination of
the lib/ directory based on arch and testdir names, but it did not
entirely clean that mess up.

Remove the arch_libdir->arch->testdir heuristic and just require
everybody sets arch_libdir correctly. Fail if the arch lib or lib asm
directories can not be found.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>