Add a MIGRATION=continuous option for stress testing purposes
Cooperative migration protocol is intrusive for the test program and
can mask some problems. Add a continuous migration mode that performs
migrations in the background for as long as the test runs.

Add a memory stress test case to selftests-migration that reproduces
a QEMU TCG memory dirty bitmap migration bug.